Fix test compilation error w/o `test` feature
